完成300

+++ schoollife/src/main/java/com/my/schoollife/controller/DeviceController.java
@@ -38,6 +38,65 @@ public class DeviceController extends BaseController {
 		return JSONObject.fromObject(param,getJsonConfig()).toString();
 	}
 	
+	@RequestMapping("/bookDevice.do")
+	@ResponseBody
+	public String bookDevice(HttpSession session,Book book) {
+		RetParam<Book> param = new RetParam<>();
+		try {
+			if(TextUtil.isEmpty(book.getUserNo())||TextUtil.isEmpty(book.getDeviceName())) {
+				throw new Exception("UserNo or DeviceName is null");
+			}
+			Book s = new Book();
+			s.setDeviceName(book.getDeviceName());
+			s.setDeviceStatus("normal");
+			List<Book> ba =  deviceService.getBookByCondition(s);
+			if(ba!=null && ba.size()>0) {
+				Book bb = new Book();
+				bb.setDeviceStatus("book");
+				bb.setDeviceName(book.getDeviceName());
+				deviceService.updateBookByCondition(bb);
+				param.setRetMsg("成功");
+				param.setRetCode(SUCCESS_CODE);
+			}else {
+				throw new Exception("设备不存在或者设备已经被抢先预约");
+			}
+		} catch (Exception e) {
+			param.setRetMsg(e.getMessage());
+			param.setRetCode(FAILED_CODE);
+		}
+		return JSONObject.fromObject(param,getJsonConfig()).toString();
+	}
+	
+	@RequestMapping("/unBookDevice.do")
+	@ResponseBody
+	public String unBookDevice(HttpSession session,Book book) {
+		RetParam<Book> param = new RetParam<>();
+		try {
+			if(TextUtil.isEmpty(book.getUserNo())||TextUtil.isEmpty(book.getDeviceName())) {
+				throw new Exception("UserNo or DeviceName is null");
+			}
+			Book s = new Book();
+			s.setDeviceName(book.getDeviceName());
+			s.setDeviceStatus("book");
+			s.setUserNo(book.getUserNo());
+			List<Book> ba =  deviceService.getBookByCondition(s);
+			if(ba!=null && ba.size()>0) {
+				Book bb = new Book();
+				bb.setDeviceStatus("normal");
+				bb.setDeviceName(book.getDeviceName());
+				deviceService.updateBookByCondition(bb);
+				param.setRetMsg("成功");
+				param.setRetCode(SUCCESS_CODE);
+			}else {
+				throw new Exception("不是您的设备，取消预约失败");
+			}
+		} catch (Exception e) {
+			param.setRetMsg(e.getMessage());
+			param.setRetCode(FAILED_CODE);
+		}
+		return JSONObject.fromObject(param,getJsonConfig()).toString();
+	}
+	
 	@RequestMapping("/updateDevice.do")
 	@ResponseBody
 	public String updateDevice(HttpSession session,Book book) {
@@ -74,6 +133,9 @@ public class DeviceController extends BaseController {
 	public String deleteDevice(HttpSession session,Book book) {
 		RetParam<Book> param = new RetParam<>();
 		try {
+			if(TextUtil.isEmpty(book.getUserNo())||TextUtil.isEmpty(book.getDeviceName())) {
+				throw new Exception("UserNo or DeviceName is null");
+			}
 			deviceService.deleteBook(book);
 			param.setRetCode(SUCCESS_CODE);
 			param.setRetMsg("删除成功");
